Moved base class for WelcomePageContainer from features/welcome/base to features/welcome
