--- features/welcome/components/base/index.js
@@ -1,1 +1,0 @@
-export * from './WelcomePageContainer';
