+++ features/welcome/components/native/WelcomePageContainer.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { Text, TextInput, TouchableHighlight, View } from 'react-native';
 
-import { WelcomePageContainer as BaseWelcomePageContainer } from '../base';
+import {
+    WelcomePageContainer as BaseWelcomePageContainer
+} from '../WelcomePageContainer';
 import { styles } from '../styles';
 
 /**
